pyproject.toml
[project]
name = "selfproxy"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "yyhhyyyyyy", email = "yyhhyyyyyy8@gmail.com"},
]
dependencies = ["requests>=2.32.4", "pyyaml>=6.0.2"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
